lib/pricing-engine/requirements.txt
psycopg2-binary
requests
boto3

